+++ tests/testthat/test-Cherries.R
@@ -1,0 +1,9 @@
+test_that("Cherries works", {
+  expect_equal(Cherries(BalancedTree(8)), 4L)
+  expect_equal(Cherries(PectinateTree(8)), 1L)
+  expect_equal(Cherries(UnrootTree(PectinateTree(8))$edge, 8L), 2L)
+  expect_error(Cherries(matrix(4, 4, 4)), "edge matrix")
+  expect_error(Cherries(1:3), "edge matrix")
+  
+  expect_no_error(Cherries(BalancedTree(144)))
+})